feat: controllers Conteudo e OptionsData
mais simplicidade e redução de arquivos
camada de controller criada para atender
as requisições, repositories agora só
fazem a consulta no banco de dados para
dar aos controllers os dados solicidados,
arquivos da pasta public/includes foram
substituídos por apenas um arquivo optionsData.php
