--- public/includes/buscarClassificacaoEtaria.php
@@ -1,9 +1,0 @@
-<?php
-
-use App\repository\BuscarClassificacaoEtaria;
-
-include_once "../../vendor/autoload.php";
-
-header('Content-Type: application/json');
-
-echo BuscarClassificacaoEtaria::getClassificacaoEtaria();
