+++ public/includes/optionsData.php
@@ -1,0 +1,7 @@
+<?php
+
+use App\controllers\OptionsData;
+
+include_once "../../vendor/autoload.php";
+
+echo OptionsData::findOptionsByName($_GET["option"]);
